Change global forms variable to a function - so default form values can't be modified. Buildbot now automatically generates getForms function change isvalid functions to use string slice so function signatures are more consistent Store form expiry date directly on form so sessionInfo struct is no longer needed change form actions integer to uint8 comment out All page

+++ !/pClubSettings.go
@@ -16,7 +16,7 @@ func clubSettings(w http.ResponseWriter, r *http.Request, clubID string) {
 	trace.Println(club.Latitude, club.Longitude)
 
 	var invalidForm, detailsForm, newMoundForm form
-	invalidForm = getSession(w, r, []int{clubDetails})
+	invalidForm = getSession(w, r, []uint8{clubDetails})
 	if invalidForm.action == clubDetails {
 		detailsForm = invalidForm
 	} else {
